🔨fix: 삭제 확인 모달 출력 안되는 버그 수정
