+++ grass-diary/src/pages/Diary/modal/UnmodifyModal.jsx
@@ -35,12 +35,15 @@ const styles = stylex.create({
     fontSize: '18px',
     fontWeight: 'bold',
   },
+  ie: {
+    fontSize: '13px',
+    marginTop: '20px',
+  },
 });
 
-const CompleteDeleteModal = ({ setShowModal }) => {
-  const closeModal = () => {
-    setShowModal(false);
-  };
+const UnmodifyModal = ({ show }) => {
+  const closeModal = () => show(false);
+
   return (
     <div {...stylex.props(styles.background)}>
       <div {...stylex.props(styles.container)}>
@@ -49,7 +52,11 @@ const CompleteDeleteModal = ({ setShowModal }) => {
         </div>
         <div {...stylex.props(styles.text)}>
           <span {...stylex.props(styles.bold)}>
-            <i className="fa-solid fa-circle-check"></i> 일기가 삭제되었습니다
+            ⚠ 수정 가능한 시간이 아닙니다
+          </span>
+          <br />
+          <span {...stylex.props(styles.ie)}>
+            일기는 당일 06:00부터 다음 날 05:59까지 수정 가능합니다.
           </span>
         </div>
       </div>
@@ -57,4 +64,4 @@ const CompleteDeleteModal = ({ setShowModal }) => {
   );
 };
 
-export default CompleteDeleteModal;
+export default UnmodifyModal;
